renamed common.py to conftest.py for pytest to recognize it
