+++ lettuce/util/__init__.py
@@ -3,26 +3,28 @@ Utility functions.
 '''
 
 from .utility import *
-# from .moments import *
+from .moments import *
 from .datautils import *
 from .utility import get_subclasses
 
-__all__ = ['get_subclasses',
-           'LettuceException',
-           'LettuceWarning',
-           'InefficientCodeWarning',
-           'ExperimentalWarning',
-           'torch_gradient',
-           'grid_fine_to_coarse',
-           'torch_jacobi',
-           'append_axes',
-           'HDF5Reporter',
-           'LettuceDataset', ]
-# 'moment_tensor',
-# 'get_default_moment_transform',
-# 'Moments',
-# 'Transform',
-# 'D1Q3Transform',
-# 'D2Q9Lallemand',
-# 'D2Q9Dellar',
-# 'D3Q27Hermite']
+__all__ = [
+    'get_subclasses',
+    'LettuceException',
+    'LettuceWarning',
+    'InefficientCodeWarning',
+    'ExperimentalWarning',
+    'torch_gradient',
+    'grid_fine_to_coarse',
+    'torch_jacobi',
+    'append_axes',
+    'HDF5Reporter',
+    'LettuceDataset',
+    'moment_tensor',
+    'get_default_moment_transform',
+    'Moments',
+    'Transform',
+    'D1Q3Transform',
+    'D2Q9Lallemand',
+    'D2Q9Dellar',
+    'D3Q27Hermite'
+]
