+++ tests/boundary/test_bounceback_bc.py
@@ -1,4 +1,4 @@
-from tests.common import *
+from tests.conftest import *
 
 from copy import copy
