+++ tests/collision/test_collision_fixpoint_2x.py
@@ -1,4 +1,4 @@
-from tests.common import *
+from tests.conftest import *
 
 
 @pytest.mark.parametrize("Collision", [BGKCollision])
