+++ tests/collision/test_collision_relaxes_shear_moments.py
@@ -1,4 +1,4 @@
-from tests.common import *
+from tests.conftest import *
 
 
 @pytest.mark.parametrize("Collision",
