+++ tests/moments/test_inverse_transform.py
@@ -1,4 +1,4 @@
-from tests.common import *
+from tests.conftest import *
 
 
 def test_inverse_transform(fix_transform):
