+++ tests/native/test_native_no_streaming_mask.py
@@ -1,4 +1,4 @@
-from tests.common import *
+from tests.conftest import *
 
 
 def test_native_no_streaming_mask():
