+++ tests/reporter/test_HDF5Reporter.py
@@ -1,5 +1,5 @@
 import os
-from tests.common import *
+from tests.conftest import *
 
 
 def test_HDF5Reporter(tmpdir):
