+++ tests/reporter/test_generic_reporters.py
@@ -1,4 +1,4 @@
-from tests.common import *
+from tests.conftest import *
 
 
 @pytest.mark.parametrize("Observable", [Enstrophy,
