+++ tests/reporter/test_write_image.py
@@ -1,4 +1,4 @@
-from tests.common import *
+from tests.conftest import *
 import os
