+++ tests/test_equilibrium.py
@@ -1,4 +1,4 @@
-from tests.common import *
+from tests.conftest import *
 
 
 @pytest.mark.parametrize("fix_equilibrium", [QuadraticEquilibrium])
