+++ tests/util/test_grid_fine_to_coarse.py
@@ -1,4 +1,4 @@
-from tests.common import *
+from tests.conftest import *
 
 
 @pytest.mark.parametrize("dims", [2, 3])
